requirements.txt
Django==1.11.17
django-filter==0.13.0
djangorestframework==3.9.4
Pillow==5.1.0
gunicorn==19.9.0
django-heroku==0.3.1
whitenoise==4.1.2
psycopg2==2.8.2
django-cors-headers==3.0.1